Lab6/pyproject.toml
[project]
name = "lab6"
version = "0.1.0"
description = "Handwritten digit recognition system with tkinter GUI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
]
